% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admb_clean.r
\name{admb_clean}
\alias{admb_clean}
\title{admb_clean}
\usage{
admb_clean(fn, which = c("sys", "output"))
}
\arguments{
\item{fn}{ADMB model name}
}
\value{
removes ADMB files
}
\description{
remove extraneous ADMB files
}
