% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admb_compile.r
\name{admb_compile}
\alias{admb_compile}
\title{admb_compile}
\usage{
admb_compile(
  fn,
  safe = FALSE,
  re = FALSE,
  verbose = FALSE,
  admb_errors = c("stop", "warn", "ignore")
)
}
\arguments{
\item{fn}{ADMB model name}
}
\value{
ADMB executable files
}
\description{
compile ADMB model
}
