% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admb_read.r
\name{admb_read}
\alias{admb_read}
\title{admb_read}
\usage{
admb_read(
  fn,
  verbose = FALSE,
  profile = FALSE,
  mcmc = FALSE,
  mcmc.opts = NULL,
  admbOut = NULL,
  checkterm = TRUE
)
}
\arguments{
\item{fn}{ADMB model name}
}
\value{
ADMB files
}
\description{
read tpl into ADMB
}
