% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admb_run.r
\name{admb_run}
\alias{admb_run}
\title{admb_run}
\usage{
admb_run(
  fn,
  verbose = FALSE,
  mcmc = FALSE,
  mcmc.opts = mcmc.control(),
  profile = FALSE,
  extra.args = "",
  admb_errors = c("stop", "warn", "ignore")
)
}
\arguments{
\item{fn}{ADMB model name}
}
\value{
ADMB files
}
\description{
run ADMB model
}
