% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_error.r
\name{calc_error}
\alias{calc_error}
\title{calc_error}
\usage{
calc_error(true, est)
}
\arguments{
\item{true}{true value (being compared to)}

\item{est}{estimate value}
}
\value{
list of each metric
}
\description{
calculate relative error, absolute error, median relative error (MRE), and median absolute relative error (MARE)
}
