% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_RTMB_model.r
\name{check_RTMB_model}
\alias{check_RTMB_model}
\title{check_RTMB_model}
\usage{
check_RTMB_model(obj, opt)
}
\arguments{
\item{obj}{objective function of RTMB model}

\item{opt}{result from nlminb optimizer}
}
\value{
list of model diagnostics (convergence, gradient, Hessian, identifiability)
}
\description{
check convergence, gradient, Hessian, and identifiability
}
