% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_admb.r
\name{clean_admb}
\alias{clean_admb}
\title{clean_admb}
\usage{
clean_admb(fn, which = c("sys", "output"))
}
\arguments{
\item{fn}{ADMB model name}
}
\value{
cleaned out local directory
}
\description{
clean extra ADMB files after compiling and running the model
}
