% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.r
\name{color_palettes}
\alias{color_palettes}
\title{color_palettes}
\usage{
color_palettes(name = NULL, reverse = FALSE, plot_colors = FALSE)
}
\arguments{
\item{name}{name of color palette. see ...}
}
\value{
vector of colors
}
\description{
provides vector of color values from palette
}
