% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_funs.r
\name{compile_admb}
\alias{compile_admb}
\title{compile_admb}
\usage{
compile_admb(
  fn,
  safe = FALSE,
  re = FALSE,
  verbose = FALSE,
  admb_errors = c("stop", "warn", "ignore")
)
}
\arguments{
\item{fn}{ADMB model name}
}
\value{
ADMB executable files
}
\description{
compile ADMB model
}
