% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_funs.r
\name{read_admb}
\alias{read_admb}
\title{read_admb}
\usage{
read_admb(
  fn,
  verbose = FALSE,
  profile = FALSE,
  mcmc = FALSE,
  mcmc.opts = NULL,
  admbOut = NULL,
  checkterm = TRUE
)
}
\arguments{
\item{fn}{ADMB model name}
}
\value{
ADMB files
}
\description{
read tpl into ADMB
}
