% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_funs.r
\name{run_admb}
\alias{run_admb}
\title{run_admb}
\usage{
run_admb(
  fn,
  verbose = FALSE,
  mcmc = FALSE,
  mcmc.opts = mcmc.control(),
  profile = FALSE,
  extra.args = "",
  admb_errors = c("stop", "warn", "ignore")
)
}
\arguments{
\item{fn}{ADMB model name}
}
\value{
ADMB files
}
\description{
run ADMB model
}
