% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci_transl.r
\name{sci_transl}
\alias{sci_transl}
\title{sci_transl}
\usage{
sci_transl(range)
}
\arguments{
\item{range}{range of values to convert scientific notation c(min, max)}
}
\value{
list of objects with scale, label, and range
}
\description{
helps do scientific notation without the e+XX stuff (created by Z. Siders)
}
