% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_lc.r
\name{theme_lc}
\alias{theme_lc}
\title{theme_lc}
\usage{
theme_lc(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{base size}

\item{base_family}{base family}
}
\description{
custom theme for ggplot (based on ggsidekik from S. Andersen)
}
